proper synthesis structure
