+++ qiskit/__init__.py
@@ -80,7 +80,7 @@ sys.modules["qiskit._accelerate.sparse_pauli_op"] = _accelerate.sparse_pauli_op
 sys.modules["qiskit._accelerate.stochastic_swap"] = _accelerate.stochastic_swap
 sys.modules["qiskit._accelerate.two_qubit_decompose"] = _accelerate.two_qubit_decompose
 sys.modules["qiskit._accelerate.vf2_layout"] = _accelerate.vf2_layout
-sys.modules["qiskit._accelerate.permutation"] = _accelerate.permutation
+sys.modules["qiskit._accelerate.synthesis.permutation"] = _accelerate.synthesis.permutation
 
 from qiskit.exceptions import QiskitError, MissingOptionalLibraryError
